Lots of changes for the fit sub-packages.
The genetic algorithm is now starting to evolve to accept templated option
paramters.  This allows us to change out the local optimizer, among other
things.  
I added a ThreadedExecutor that is to be used with APPSPACK.
I also added an implementation of a local optimizer for the genetic algorithm
that can call APPSPACK.